Added RenderTarget scaling to mouse events
Scaling now works for mouse coordinates. This means that if you resize
the window without changing the sf::View size of the RenderTarget your
mouse events will still be at the correct place in terms of X, Y. Since
Guichan doesn't "know" about the SFML view the coordinates have to be
normalized to "Guichan" space.